final fix

+++ client/src/components/TrackDetails.jsx
@@ -1,11 +1,20 @@
 import React from "react";
 import { useTracksContext } from "../hooks/useTracksContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 function TrackDetails({ track }) {
+  const { user } = useAuthContext();
   const { dispatch } = useTracksContext();
   const handleClick = async () => {
-    const response = await fetch(`http://localhost:4001/tracks/${track._id}`, {
+    if (!user) {
+      return;
+    }
+
+    const response = await fetch("http://localhost:4001/tracks/" + track._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
